+++ src/Render/NeoInputRenderer.php
@@ -3,7 +3,7 @@
 namespace Efabrica\NeoForms\Render;
 
 use Efabrica\NeoForms\Control\ToggleSwitch;
-use Efabrica\NeoForms\NeoFormRenderer;
+use Efabrica\NeoForms\Render\NeoFormRenderer;
 use Efabrica\Nette\Forms\Rte\RteControl;
 use Nette\Forms\Controls\BaseControl;
 use Nette\Forms\Controls\Button;
@@ -13,6 +13,7 @@ use Nette\Forms\Controls\MultiSelectBox;
 use Nette\Forms\Controls\SelectBox;
 use Nette\Forms\Controls\TextArea;
 use Nette\Forms\Controls\TextInput;
+use Nette\Forms\Controls\UploadControl;
 use Nette\Utils\Html;
 use RadekDostal\NetteComponents\DateTimePicker\AbstractDateTimePicker;
 use RadekDostal\NetteComponents\DateTimePicker\DateTimePicker;
@@ -22,10 +23,12 @@ use RuntimeException;
 class NeoInputRenderer
 {
     private NeoFormRenderer $renderer;
+    private NeoInputViewRenderer $viewRenderer;
 
     public function __construct(NeoFormRenderer $renderer)
     {
         $this->renderer = $renderer;
+        $this->viewRenderer = new NeoInputViewRenderer($renderer);
     }
 
     private function block(string $blockName, array $attrs): string
@@ -35,12 +38,15 @@ class NeoInputRenderer
 
     public function input(BaseControl $el, array $options = []): string
     {
+        /** @var Html $control */
+        $control = $el->getControl();
+        if (($options['readonly'] ?? false) || $el->getOption('readonly')) {
+            return $this->viewRenderer->input($el);
+        }
         if ($el instanceof Checkbox) {
             return $this->checkbox($el, $options);
         }
 
-        /** @var Html $control */
-        $control = $el->getControl();
         $attrs = $control->attrs;
         unset($attrs['data-nette-rules']);
         $attrs += array_filter($options, 'is_scalar');
@@ -48,8 +54,6 @@ class NeoInputRenderer
         $s = '';
         if ($el instanceof AbstractDateTimePicker) {
             $s .= $this->datepicker($el, $attrs, $options);
-        } elseif ($el instanceof TextInput) {
-            $s .= $this->textInput($el, $attrs, $options);
         } elseif ($el instanceof SelectBox || $el instanceof MultiSelectBox) {
             $s .= $this->selectBox($el, $attrs, $options);
         } elseif ($el instanceof Button) {
@@ -58,14 +62,16 @@ class NeoInputRenderer
             $s .= $this->textarea($el, $attrs, $options);
         } elseif ($el instanceof HiddenField) {
             $s .= $this->hidden($el, $attrs, $options);
+        } elseif ($el instanceof UploadControl) {
+            $s .= $this->upload($el, $attrs, $options);
         } else {
-            throw new RuntimeException(get_class($el) . ' is not yet supported in NeoFormRenderer');
+            $s .= $this->textInput($el, $attrs, $options);
         }
 
         return $s . $this->description($el, $options);
     }
 
-    public function textInput(TextInput $el, array $attrs, array $options): string
+    public function textInput(BaseControl $el, array $attrs, array $options): string
     {
         return $this->block('inputText', [
             'attrs' => $attrs,
@@ -96,9 +102,9 @@ class NeoInputRenderer
 
     public function checkbox(Checkbox $checkbox, array $options): string
     {
-
+        $caption = $options['caption'] ?? '';
         return $this->block($checkbox instanceof ToggleSwitch ? 'toggleSwitch' : 'checkbox', [
-            'caption' => $checkbox->getCaption(),
+            'caption' => $caption === true ? $checkbox->getCaption() : $caption,
             'info' => $checkbox->getOption('info'),
             'labelAttrs' => array_filter($checkbox->getLabelPart()->attrs, 'is_scalar'),
             'inputAttrs' => array_filter($checkbox->getControlPart()->attrs, 'is_scalar'),
@@ -139,6 +145,15 @@ class NeoInputRenderer
             'el' => $el,
             'description' => $el->getOption('description'),
             'options' => $options,
+        ]);
+    }
+
+    public function upload(UploadControl $el, array $attrs, array $options): string
+    {
+        return $this->block('upload', [
+            'attrs' => $attrs,
+            'options' => $el->getOptions() + $options,
+            'description' => $el->getOption('description'),
         ]);
     }
 }